requirements.txt
aniso8601==4.0.1
astroid==2.1.0
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4.3
bcrypt==3.1.4
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
coverage==4.5.2
coveralls==1.5.1
docopt==0.6.2
flake8==3.6.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-RESTful==0.3.6
gunicorn==19.9.0
idna==2.7
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
more-itertools==4.3.0
passlib==1.7.1
pluggy==0.8.0
psycopg2-binary==2.7.6.1
py==1.7.0
pycodestyle==2.4.0
pycparser==2.19
pyflakes==2.0.0
PyJWT==1.7.0
pylint==2.2.2
pytest==4.0.1
pytest-cov==2.6.0
python-coveralls==2.9.1
pytz==2018.7
PyYAML==3.13
requests==2.20.1
six==1.11.0
typed-ast==1.1.0
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.10.11
